This PR was squashed before being merged into the 2.13.x-dev branch.

Discussion
----------

This Pull Request fixes the API calls for the functions:
- \Github\Api\Organization\Teams::show
- \Github\Api\Organization\Teams::update
- \Github\Api\Organization\Teams::remove
- \Github\Api\Organization\Teams::members
- \Github\Api\Organization\Teams::check
- \Github\Api\Organization\Teams::addMember
- \Github\Api\Organization\Teams::removeMember

based on the official Github's v3 API documentation:
- https://developer.github.com/v3/teams/
- https://developer.github.com/v3/teams/members/

since those functions incorrectly threw `Github\Exception\RuntimeException: Not Found`

Commits
-------

624da985f19883e5a64f554f1fc06f2a2e5c43c2 Ignore PHPStorm's .idea directory
3f332edcd4014cccdd9c5b8500c84564dc5dda53 Fix various Teams' URIs
7742bd1d322b4129ac56ece532ae2c45cd9d6738 Revert "Ignore PHPStorm's .idea directory"
ca63e29962c4a11d9e3573e2d8b826127e234aec fixes backward compatibility check
9f5071cbf7832adcd652d23fc1a69d57d0fbf51a fix update()
6ecdb0a78f48e4b4b7065b8c7a6a03eec8747261 style fix
6ceeb55dddafb90367db36b9a875c3b86ce36c4c changes based on @acrobat's suggestions
25504440b6b2fa6b6800c7c08741e55e13848a41 styling fix

+++ lib/Github/Api/Organization/Teams.php
@@ -32,12 +32,24 @@ class Teams extends AbstractApi
         return $this->post('/orgs/'.rawurlencode($organization).'/teams', $params);
     }
 
-    public function show($team)
+    /**
+     * @link https://developer.github.com/v3/teams/#list-teams
+     */
+    public function show($team, $organization = null)
     {
-        return $this->get('/teams/'.rawurlencode($team));
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->get('/teams/'.rawurlencode($team));
+        }
+
+        return $this->get('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team));
     }
 
-    public function update($team, array $params)
+    /**
+     * @link https://developer.github.com/v3/teams/#edit-team
+     */
+    public function update($team, array $params, $organization = null)
     {
         if (!isset($params['name'])) {
             throw new MissingArgumentException('name');
@@ -46,32 +58,83 @@ class Teams extends AbstractApi
             $params['permission'] = 'pull';
         }
 
-        return $this->patch('/teams/'.rawurlencode($team), $params);
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->patch('/teams/'.rawurlencode($team), $params);
+        }
+
+        return $this->patch('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team), $params);
     }
 
-    public function remove($team)
+    /**
+     * @link https://developer.github.com/v3/teams/#delete-team
+     */
+    public function remove($team, $organization = null)
     {
-        return $this->delete('/teams/'.rawurlencode($team));
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->delete('/teams/'.rawurlencode($team));
+        }
+
+        return $this->delete('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team));
     }
 
-    public function members($team)
+    /**
+     * @link https://developer.github.com/v3/teams/members/#list-team-members
+     */
+    public function members($team, $organization = null)
     {
-        return $this->get('/teams/'.rawurlencode($team).'/members');
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->get('/teams/'.rawurlencode($team).'/members');
+        }
+
+        return $this->get('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team).'/members');
     }
 
-    public function check($team, $username)
+    /**
+     * @link https://developer.github.com/v3/teams/members/#get-team-membership
+     */
+    public function check($team, $username, $organization = null)
     {
-        return $this->get('/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->get('/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
+        }
+
+        return $this->get('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
     }
 
-    public function addMember($team, $username)
+    /**
+     * @link https://developer.github.com/v3/teams/members/#add-or-update-team-membership
+     */
+    public function addMember($team, $username, $organization = null)
     {
-        return $this->put('/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->put('/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
+        }
+
+        return $this->put('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
     }
 
-    public function removeMember($team, $username)
+    /**
+     * @link https://developer.github.com/v3/teams/members/#remove-team-membership
+     */
+    public function removeMember($team, $username, $organization = null)
     {
-        return $this->delete('/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
+        if (null === $organization) {
+            @trigger_error('Not passing the $organisation parameter is deprecated in knpLabs/php-github-api v2.14 and will be mandatory in v3.0.', E_USER_DEPRECATED);
+
+            return $this->delete('/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
+        }
+
+        return $this->delete('/orgs/'.rawurlencode($organization).'/teams/'.rawurlencode($team).'/memberships/'.rawurlencode($username));
     }
 
     public function repositories($team)
